% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astar.R
\docType{data}
\name{AStar}
\alias{AStar}
\title{Convenience class for running A* algorithm.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AStar
}
\description{
Has methods for all function arguments of astar.
}
\examples{
 nodes <- list(
   A = c(B = 100, C = 20),
   C = c(D = 20),
   D = c(B = 20)
 )
 Searcher <- R6::R6Class(
   "Searcher",
   inherit = AStar,
   public = list(
     neighbors = function(node) names(nodes[[node]]),
     cost_estimate = function(node, goal) 1,
     edge_distance = function(src, dst) nodes[[src]][dst],
     is_goal_reached = function(node, goal) identical(node, goal)
   )
 )

 searcher <- Searcher$new()
 path <- searcher$run('A', 'B')

}
\keyword{datasets}
