% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{MazeGazer}
\alias{MazeGazer}
\title{Class to help visualize the inner workings of SearchMaze.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MazeGazer
}
\description{
MazeGazer adds a history entry that collects each paze the algorithm visits.
}
\examples{

M <- matrix(ncol = 4, byrow = TRUE, c(
  0, 1, 0, 0,
  0, 1, 0, 1,
  0, 1, 0, 1,
  0, 0, 0, 0
  ))

mg <- MazeGazer$new(M)
mg$run(c(1, 1), c(1, 4))

mg$history


}
\keyword{datasets}
