% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{SearchMaze2D}
\alias{SearchMaze2D}
\title{AStar subclass demonstrating 2D maze navigation.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SearchMaze2D
}
\description{
Has methods for all function arguments of astar.
Takes a 2d matrix with 0 for walkable tiles, and any other number for walls
}
\examples{
M <- matrix(ncol = 4, byrow = TRUE, c(
  0, 1, 0, 0,
  0, 1, 0, 1,
  0, 1, 0, 1,
  0, 0, 0, 0
  ))

sm <- SearchMaze2D$new(M)
sm$run(c(1, 1), c(1, 4))

}
\keyword{datasets}
